--- Library/inc/DHEngine/DH3DEngineDefine.h
@@ -1,37 +1,0 @@
-#pragma once
-
-/// 사용자 화면에 대한 기본적인 정의.
-#define SCREEN_WIDTH 1920
-#define SCREEN_HEIGHT 1080
-
-/// Release 등의 관련 Define
-#define ReleaseCOM(x) { if(x){ x->Release(); x = 0; } }
-#define SAFE_RELEASE(x) { if(x){ x->Release(); x = 0; } }
-#define SafeDelete(x) { delete x; x = 0; }
-
-#if defined(DEBUG) | defined(_DEBUG)
-#ifndef HR
-#define HR(x)									\
-	{									\
-		HRESULT hr = (x);						\
-		if(FAILED(hr))							\
-		{								\
-			LPWSTR output;                                    	\
-			FormatMessage(FORMAT_MESSAGE_FROM_SYSTEM |		\
-				FORMAT_MESSAGE_IGNORE_INSERTS 	 |		\
-				FORMAT_MESSAGE_ALLOCATE_BUFFER,			\
-				NULL,						\
-				hr,						\
-				MAKELANGID(LANG_NEUTRAL, SUBLANG_DEFAULT),	\
-				(LPTSTR) &output,				\
-				0,						\
-				NULL);					        \
-			MessageBox(NULL, output, L"Error", MB_OK);		\
-		}								\
-	}
-#endif
-#else
-#ifndef HR
-#define HR(x) (x)
-#endif
-#endif
